requirements.txt
neo4j==5.18.0
google-cloud-bigquery==3.17.2
pandas==2.2.1
tqdm==4.66.2
python-dotenv==1.0.1
pyarrow==15.0.2
nltk==3.8.1